+++ jdaviz/configs/default/plugins/export/export.py
@@ -11,6 +11,7 @@ from regions import CircleSkyRegion, EllipseSkyRegion
 from specutils import Spectrum
 from traitlets import Bool, List, Unicode, observe
 
+from jdaviz.configs.default.plugins.export.avm import png_embed_avm
 from jdaviz.core.custom_traitlets import FloatHandleEmpty, IntHandleEmpty
 from jdaviz.core.marks import ShadowMixin
 from jdaviz.core.registries import tray_registry
@@ -140,7 +141,7 @@ class Export(PluginTemplateMixin, ViewerSelectMixin, SubsetSelectMixin,
 
         self.viewer.add_filter('is_not_empty')
 
-        viewer_format_options = ['png', 'svg']
+        viewer_format_options = ['png', 'svg', 'jpg']
         if self.config == 'cubeviz':
             if not self._app.state.settings.get('server_is_remote'):
                 viewer_format_options += ['mp4']
@@ -446,7 +447,7 @@ class Export(PluginTemplateMixin, ViewerSelectMixin, SubsetSelectMixin,
 
     @with_spinner()
     def export(self, filename=None, show_dialog=None, overwrite=False,
-               raise_error_for_overwrite=True):
+               raise_error_for_overwrite=True, embed_avm=True):
         """
         Export selected item(s)
 
@@ -465,6 +466,10 @@ class Export(PluginTemplateMixin, ViewerSelectMixin, SubsetSelectMixin,
             If `True`, raise exception when ``overwrite=False`` but
             output file already exists. Otherwise, a message will be sent
             to application snackbar instead.
+
+        embed_avm : bool
+            If `True` and the file type is PNG, embed
+            Astronomy Visualization Metadata (including WCS) using ``pyAVM``.
         """
         if self.multiselect:
             raise NotImplementedError("batch export not yet supported")
@@ -504,10 +509,42 @@ class Export(PluginTemplateMixin, ViewerSelectMixin, SubsetSelectMixin,
                     mark.labels = [lbl.strip() for lbl in mark.labels]
                 restores.append(restore)
 
+            # avoid circular import:
+            from jdaviz.configs.imviz.plugins.viewers import ImvizImageView
+
             if filetype == "mp4":
                 self.save_movie(viewer, filename, filetype,
                                 width=f"{self.image_width}px" if self.image_custom_size else None,
                                 height=f"{self.image_height}px" if self.image_custom_size else None)
+
+            elif (
+                    isinstance(self.viewer.selected_obj, ImvizImageView) and
+                    embed_avm and filetype in ('jpg', 'png')
+            ):
+                # export screenshot to JPG with AVM by:
+                #   (1) export a temporary PNG
+                #   (2) convert temporary PNG to a temporary JPG with PIL
+                #   (3) use pyAVM to embed AVM into a final copy of the temporary JPG
+
+                # export as PNG, even if final filetype will be JPG
+                # note: following `replace` assumes lower case extensions
+                tmp_filename = Path(
+                    str(Path(filename)).replace('.jpg', '.png')
+                )
+
+                self.save_figure(
+                    viewer, tmp_filename, 'png', show_dialog=show_dialog,
+                    width=f"{self.image_width}px" if self.image_custom_size else None,
+                    height=f"{self.image_height}px" if self.image_custom_size else None
+                )
+
+                # wait for PNG to be available before continuing
+                while viewer.figure._upload_png_callback is not None:
+                    time.sleep(0.05)
+
+                # now convert to JPG with AVM
+                png_embed_avm(self.app._jdaviz_helper, viewer, tmp_filename, format=filetype)
+
             else:
                 self.save_figure(viewer, filename, filetype, show_dialog=show_dialog,
                                  width=f"{self.image_width}px" if self.image_custom_size else None,
@@ -633,11 +670,13 @@ class Export(PluginTemplateMixin, ViewerSelectMixin, SubsetSelectMixin,
         else:
             app = viewer.app
 
-        def on_img_received(data):
+        def on_img_received(data, filename=filename):
             try:
                 with filename.open(mode='bw') as f:
                     f.write(data)
+
             except Exception as e:
+                print(e)
                 self.hub.broadcast(SnackbarMessage(
                     f"{self.viewer.selected} failed to export to {str(filename)}: {e}",
                     sender=self, color="error", traceback=e))
@@ -646,7 +685,7 @@ class Export(PluginTemplateMixin, ViewerSelectMixin, SubsetSelectMixin,
                     f"{self.viewer.selected} exported to {str(filename)}",
                     sender=self, color="success"))
 
-        def get_png(figure):
+        def get_png(figure, on_img_received=on_img_received):
             if figure._upload_png_callback is not None:
                 raise ValueError("previous png export is still in progress. Wait to complete before making another call to save_figure")  # noqa: E501 # pragma: no cover
 
@@ -701,9 +740,10 @@ class Export(PluginTemplateMixin, ViewerSelectMixin, SubsetSelectMixin,
                 threading.Thread(target=wait_in_other_thread).start()
             _widget_after_first_display(cloned_viewer.figure, on_figure_displayed)
             _show_hidden(cloned_viewer.figure, width, height)
-        elif filetype == 'png':
+        elif filetype in ('png', 'jpg'):
             # NOTE: get_png already check if _upload_png_callback is not None
             get_png(viewer.figure)
+
         elif filetype == 'svg':
             if viewer.figure._upload_svg_callback is not None:
                 raise ValueError("previous svg export is still in progress. Wait to complete before making another call to save_figure") # noqa